+++ CHANGELOG.md
@@ -75,6 +75,7 @@ here: [Community Features](https://github.com/SynthstromAudible/DelugeFirmware/b
 - Added a feature save user-defined pad brightness level and restore it at startup.
 - Mod (Gold) Encoder LED indicators are now Bipolar for Bipolar params (e.g. Param, Pitch, Patch Cables). Positive values illuminate the top two LEDs. Negative values illuminate the bottom two LEDs. The middle value doesn't light up any LEDs. 
 - The play button now blinks when the the CPU Usage of the Deluge is high and synth voices/sample playback are being culled.
+- Removed ability to convert an Audio Clip to an Instrument Clip (Synth / Kit / MIDI / CV) as this conversion process is error/bug prone.
 
 In addition, a number of improvements have been made to how the OLED display is used: